♻️ (install_asdf.sh): refactor script to include user prompt and improve installation process 🔧 (setup.sh): enable asdf installation script in setup process

+++ setup/scripts/install_asdf.sh
@@ -1,25 +1,21 @@
-#!/bin/bash
+if ask "Do you want to install asdf?"; then
+  ARCH=$(uname -m)
+  [[ "$ARCH" == "x86_64" ]] && ARCH=amd64
+  [[ "$ARCH" == "aarch64" ]] && ARCH=arm64
 
-BOLD='\033[1m'
-GREEN='\033[32m'
-RESET='\033[0m'
+  LATEST_URL=$(curl -s https://api.github.com/repos/asdf-vm/asdf/releases/latest \
+    | jq -r --arg arch "$ARCH" '.assets[] | select(.name | test("asdf.*linux-" + $arch + ".*\\.tar\\.gz$")).browser_download_url' \
+    | head -n 1)
 
-ARCH=$(uname -m)
-[[ "$ARCH" == "x86_64" ]] && ARCH=amd64
-[[ "$ARCH" == "aarch64" ]] && ARCH=arm64
+  curl -L -o asdf.tar.gz $LATEST_URL
 
-LATEST_URL=$(curl -s https://api.github.com/repos/asdf-vm/asdf/releases/latest \
-  | jq -r --arg arch "$ARCH" '.assets[] | select(.name | test("asdf.*linux-" + $arch + ".*\\.tar\\.gz$")).browser_download_url' \
-  | head -n 1)
+  sudo rm -rf $HOME/.local/bin/asdf
 
-curl -L -o asdf.tar.gz $LATEST_URL
+  sudo mkdir -p $HOME/.local/bin
 
-rm -rf /home/ton618/.local/bin/asdf
+  sudo tar -xzf asdf.tar.gz -C /home/ton618/.local/bin
 
-mkdir -p /home/ton618/.local/bin
+  rm asdf.tar.gz
 
-tar -xzf asdf.tar.gz -C /home/ton618/.local/bin
-
-rm asdf.tar.gz
-
-echo -e "${BOLD}${GREEN}asdf is installed.${RESET}"
+  success "asdf installed!"
+fi
